Finished Uninstaller

+++ Uninstaller.py
@@ -15,31 +15,36 @@ class Uninstaller:
         # Defines the parser for the cfg files
         self.cfg_parser = configparser.RawConfigParser()
 
+        # Defines the popup
+        self.popup = ""
+
         # Sets the paths
         self.paths = []
         self.install_path = ""
         self.autostart_path = ""
+        self.autostart_file_path = ""
 
         # Gets the data from the cfg file
-        self.cfg_path = os.path.abspath("config.cfg")
-        self.get_data()
+        if os.path.isfile("config.cfg"):
+            self.cfg_path = os.path.abspath("config.cfg")
+            self.get_data()
 
-        # Runs the popup
-        self.window = ""
-        self.popup = ""
-        self.run_popup()
+            # Runs the popup
+            self.run_popup()
+
+        else:
+            sg.PopupError("No install config found!")
 
     def run_popup(self):
         """Shows and runs the popup"""
         self.popup = sg.PopupYesNo("Uninstall AWC Tool?")
 
         if self.popup == "Yes":
-            print("hello")
             # Uninstalls all files
             self.uninstall()
 
         elif (self.popup == "No") or (self.popup == sg.WIN_CLOSED):
-            print("bye")
+            sg.Popup("Uninstall Canceled!")
 
     def get_data(self):
         """Gets the data of all files"""
@@ -47,16 +52,25 @@ class Uninstaller:
         self.install_path = self.cfg_parser.get("Install-Config", "Install_Path")
         self.autostart_path = self.cfg_parser.get("Install-Config", "Autostart_Path")
 
+        self.autostart_file_path = os.path.join(self.autostart_path, "AWC.lnk")
+
     def uninstall(self):
         """Uninstalls all files"""
         sg.Popup("All Files Removed!")
 
         # Terminates the process
-        os.system("taskkill /F /IM AWC.exe")
+        try:
+            os.system("taskkill /F /IM AWC.exe")
+
+        except Exception as e:
+            pass
 
         # Removes all files
-        os.remove(self.autostart_path)
-        shutil.rmtree(self.install_path)
+        if os.path.isfile(self.autostart_file_path):
+            os.remove(self.autostart_path)
+
+        if os.path.isdir(self.install_path):
+            shutil.rmtree(self.install_path)
 
 
 if __name__ == "__main__":
